Release adaptive runtime and secret-managed deploy

+++ mobile/asystence_mobile/lib/src/features/workspace/huddle_media/livekit_huddle_media_provider.dart
@@ -1249,7 +1249,7 @@ class LiveKitHuddleMediaProvider extends HuddleMediaProvider {
             maxFramerate: maxFramerate,
           ),
         ),
-        maxFrameRate: maxFramerate,
+        maxFrameRate: maxFramerate.toDouble(),
       ),
       defaultVideoPublishOptions: lk.VideoPublishOptions(
         videoEncoding: lk.VideoEncoding(
